refactor: moved VideoBuffer UserControl
