+++ src/Spice86/ViewModels/PaletteViewModel.cs
@@ -10,7 +10,6 @@ using CommunityToolkit.Mvvm.ComponentModel;
 
 using Spice86.Core.Emulator.Devices.Video;
 using Spice86.Core.Emulator.VM;
-using Spice86.Shared;
 using Spice86.Shared.Emulator.Video;
 
 public partial class PaletteViewModel : ObservableObject {
