+++ src/Spice86/Views/PerformanceWindow.axaml.cs
@@ -1,4 +1,5 @@
 namespace Spice86.Views;
+
 using Avalonia;
 using Avalonia.Controls;
 using Avalonia.Controls.ApplicationLifetimes;
