--- src/Spice86/Views/VideoBufferView.axaml
@@ -1,25 +1,0 @@
-<UserControl xmlns="https://github.com/avaloniaui"
-			 xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-			 xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-			 xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-			 xmlns:vm="using:Spice86.ViewModels"
-			 mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-			 x:Class="Spice86.Views.VideoBufferView">
-	<Design.DataContext>
-		<vm:VideoBufferViewModel />
-	</Design.DataContext>
-	<LayoutTransformControl>
-		<LayoutTransformControl.RenderTransform>
-			<ScaleTransform
-				ScaleX="{Binding Scale}"
-				ScaleY="{Binding Scale}">
-			</ScaleTransform>
-		</LayoutTransformControl.RenderTransform>
-		<Viewbox>
-			<Image x:Name="Image"
-				RenderOptions.BitmapInterpolationMode="None"
-				Cursor="{Binding Cursor}"
-				Source="{Binding Bitmap}" />
-		</Viewbox>
-	</LayoutTransformControl>
-</UserControl>
